ios: unify filter and sort between forward and share sheets
